Fix getColorHex method in Label Model
The bug was made from casting color.getRed(), color.getGreen(),
color.getBlue() to int which always lead to approximating a
less-than-zero-value to zero, then multiplying it by 255 which gives
a zero as a final result, so the result of the formatted string would
always be "#000000".

This bug was fixed by casting to int after the multiplication is done.